cloudrun/requirements.txt
alembic==1.9.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
CacheControl==0.14.0
docstring_parser==0.16
fastapi==0.103.2
firebase-admin==6.5.0
flatbuffers==24.3.25
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-cloud-pubsub==2.21.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
#tensorflow
tensorflow==2.12.0
tensorflow-io==0.32.0
numpy~=1.23.5
pillow~=10.4.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
jax==0.4.13
# jaxlib==0.4.30
libclang==18.1.1
Mako==1.3.5
markdown-it-py==3.0.0
# ml-dtypes==0.3.2
ml-dtypes==0.2.0
msgpack==1.0.8
optree==0.11.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests-oauthlib==1.3.0
rich==13.7.1
shapely==2.0.4
SQLAlchemy==1.4.46
starlette==0.27.0
tensorboard==2.12.3
tensorboard-plugin-wit==1.8.1
# tensorflow-io-gcs-filesystem==0.37.1
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.4.0
uritemplate==4.1.1
uvicorn==0.23.2
uvloop==0.19.0
vertexai==1.49.0
watchfiles==0.22.0
websockets==12.0
sniffio==1.3.1
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
h5py==3.8.0
loguru==0.7.2
pydantic-settings==2.3.4
babel==2.15.0
playwright==1.45.0
cryptography==43.0.0
functions-framework==3.*
cloudevents==1.11.0